Remove account contract module
