+++ modules/minimint-simp/src/contracts/mod.rs
@@ -1,4 +1,3 @@
-pub mod account;
 pub mod incoming;
 pub mod outgoing;
 
@@ -28,7 +27,7 @@ hash_newtype!(
 #[allow(clippy::large_enum_variant)]
 #[derive(Debug, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, Encodable, Decodable)]
 pub enum Contract {
-    Account(account::AccountContract),
+    // Account(account::AccountContract),
     Incoming(incoming::IncomingContract),
     Outgoing(outgoing::OutgoingContract),
 }
@@ -37,7 +36,7 @@ pub enum Contract {
 #[allow(clippy::large_enum_variant)]
 #[derive(Debug, Clone, Eq, PartialEq, Hash, Encodable, Decodable, Serialize, Deserialize)]
 pub enum FundedContract {
-    Account(account::AccountContract),
+    // Account(account::AccountContract),
     Incoming(incoming::FundedIncomingContract),
     Outgoing(outgoing::OutgoingContract),
 }
@@ -54,7 +53,7 @@ pub enum ContractOutcome {
 impl IdentifyableContract for Contract {
     fn contract_id(&self) -> ContractId {
         match self {
-            Contract::Account(c) => c.contract_id(),
+            // Contract::Account(c) => c.contract_id(),
             Contract::Incoming(c) => c.contract_id(),
             Contract::Outgoing(c) => c.contract_id(),
         }
@@ -64,7 +63,7 @@ impl IdentifyableContract for Contract {
 impl IdentifyableContract for FundedContract {
     fn contract_id(&self) -> ContractId {
         match self {
-            FundedContract::Account(c) => c.contract_id(),
+            // FundedContract::Account(c) => c.contract_id(),
             FundedContract::Incoming(c) => c.contract.contract_id(),
             FundedContract::Outgoing(c) => c.contract_id(),
         }
@@ -76,7 +75,7 @@ impl Contract {
     /// the contract type it is not yet final.
     pub fn to_outcome(&self) -> ContractOutcome {
         match self {
-            Contract::Account(_) => ContractOutcome::Account,
+            // Contract::Account(_) => ContractOutcome::Account,
             Contract::Incoming(_) => ContractOutcome::Incoming(DecryptedPreimage::Pending),
             Contract::Outgoing(_) => ContractOutcome::Outgoing,
         }
@@ -85,7 +84,7 @@ impl Contract {
     /// Converts a contract to its executed version.
     pub fn to_funded(self, out_point: OutPoint) -> FundedContract {
         match self {
-            Contract::Account(account) => FundedContract::Account(account),
+            // Contract::Account(account) => FundedContract::Account(account),
             Contract::Incoming(incoming) => FundedContract::Incoming(FundedIncomingContract {
                 contract: incoming,
                 out_point,
